Remove more PHP 7.0 compatibility code
