+++ phpunit.xml.dist
@@ -9,9 +9,6 @@
   <filter>
     <whitelist processUncoveredFilesFromWhitelist="true">
       <directory suffix=".php">src/</directory>
-      <exclude>
-        <file>src/Application/CoversValidator.php</file>
-      </exclude>
     </whitelist>
   </filter>
 </phpunit>
