+++ src/Application/CoversValidator.php
@@ -2,13 +2,51 @@
 
 namespace OckCyp\CoversValidator\Application;
 
-use Composer\InstalledVersions;
+use OckCyp\CoversValidator\Command\ValidateCommand;
+use Symfony\Component\Console\Application;
+use Symfony\Component\Console\Input\InputDefinition;
+use Symfony\Component\Console\Input\InputInterface;
 
-$consoleVersion = InstalledVersions::getVersion('symfony/console');
+class CoversValidator extends Application
+{
+    public const NAME = 'CoversValidator';
+    public const VERSION = '1.5.0';
 
-$chosenClass = CoversValidatorSymfony5::class;
-if (\version_compare($consoleVersion, '6.0.0') >= 0) {
-    $chosenClass = CoversValidatorSymfony6::class;
+    /**
+     * {@inheritdoc}
+     */
+    public function __construct($name = self::NAME, $version = self::VERSION)
+    {
+        parent::__construct($name, $version);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getCommandName(InputInterface $input): ?string
+    {
+        return 'validate';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function getDefaultCommands(): array
+    {
+        $defaultCommands = parent::getDefaultCommands();
+        $defaultCommands[] = new ValidateCommand();
+
+        return $defaultCommands;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDefinition(): InputDefinition
+    {
+        $inputDefinition = parent::getDefinition();
+        $inputDefinition->setArguments();
+
+        return $inputDefinition;
+    }
 }
-
-\class_alias($chosenClass, CoversValidator::class);
